backend/requirements.txt
psycopg2-binary
fastapi
starlette
urllib3
pydantic